requirements.txt
fastapi
uvicorn[standard]
gunicorn
botbuilder-core
botbuilder-schema
botbuilder-integration-aiohttp
python-multipart
